api/requirements.txt
Django==4.0
djangorestframework==3.13.1
django-cors-headers==3.11.0
redis==4.1.0
psycopg2==2.9.1